requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
